Add click event to menuitem + minor refact
