--- components/MenuItem.vue
@@ -1,55 +1,0 @@
-<template>
-  <div class="nav-item dropdown" v-if="menuItem.type === 'submenu'" >
-    <a
-        :id="menuItem.id"
-        ref="buttonRef"
-        @click="dropdown.show()"
-        class="nav-link dropdown-toggle"
-        role="button"
-        data-bs-toggle="dropdown"
-        aria-haspopup="true"
-        aria-expanded="false"
-    >
-      {{ menuItem.name }}
-    </a>
-    <ul class="dropdown-menu"
-        ref="menuRef"
-        :aria-labelledby="menuItem.id">
-      <li
-          v-for="(m, i) in menuItem.submenu"
-          :key="i"
-      >
-        <a
-            v-if="m.type === 'item'"
-            class="dropdown-item"
-            href="#"
-        >
-          {{ m.name }}
-        </a>
-      </li>
-    </ul>
-  </div>
-</template>
-
-<script setup lang="ts">
-  import { ref, onMounted } from 'vue'
-  import { IMenuItem } from "~/store/appData";
-  import {Dropdown} from "bootstrap";
-  const { $bootstrap } = useNuxtApp()
-  let dropdown: Dropdown | null = null;
-
-    const props = defineProps<{
-      menuItem: IMenuItem;
-    }>();
-
-    const buttonRef = ref<HTMLElement | null>(null)
-
-    onMounted(() => {
-      // Ensure that the Bootstrap 5 dropdown works properly
-      dropdown = new $bootstrap.Dropdown(buttonRef.value);
-    })
-</script>
-
-<style>
-/* Add any custom styling here */
-</style>
